refactor(_api): extract _capture_colorbar_geometry to _save_capture (line budget)
_save_helpers.py reached 523 lines (>512 audit budget) once the colorbar
geometry capture landed. Move _capture_colorbar_geometry into a new
_api/_save_capture.py (save-time live-geometry capture); _save.py imports it
from there. Behaviour-neutral: verbatim function move, runtime import verified,
colorbar reproducer test passes in isolation (parity with pre-move dd966b8).

_save_helpers.py 523 -> 477, _save.py 481, _save_capture.py 58 (all < 512).
